+++ src/main/java/com/travelapp/travelapp/restcontroller/CollageController.java
@@ -33,7 +33,7 @@ public class CollageController {
     }
 
     @GetMapping("/byId/{id}")
-    public CollageDTOGet getCollagePost(@PathVariable int id){
+    public CollageDTOGet getCollagePost(@PathVariable long id){
         CollageDTOGet collage = collageService.getCollageById(id);
         return collage;
     }
